SpeechRecognition: Fix wrong state transitions
Do not allow transition from ListeningStoppingState to either
ListeningState or MutedState. Also move straight to IdleState
on abortListening() or reset() when possible.

Details:

If listening is stopped immediately after starting it, the
expected state sequence is:

IdleState
-> ListeningStartingState
-> ListeningStoppingState
-> IdleState

In this case signal listeningStarted() is now emitted on
ListeningStoppingState (was: ListeningState).

If listening was started, stopped and aborted in a fast sequence,
signal listeningStopped() was emitted even when listeningStarted()
was not. The expected behavior for abortListening() or reset() is
to move straight to IdleState in this case.

Change-Id: I3d3b9d096ad38c1bdf2f5ef7958776f2dd4a1496
Reviewed-by: Andrew Knight <andrew.knight@intopalo.com>
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@theqtcompany.com>