Revert "Fix compilation on Debian"
Debian seems to have a packaging bug, or libspeechd is just buggy,
whichever it is, the deb packages install the libspeechd.h header twice.
That header has a "#include <libspeechd_version.h>" in some versions,
but the version file is only to be found in the speech-dispatcher
subdirectory. See for example
https://packages.debian.org/sid/amd64/libspeechd-dev/filelist
which seems to be the same in all versions I could find.

This reverts commit 45a127b37ef743a4a6dade41e26bf582e96c7938.

Change-Id: If51ec7b47020d50cb95582487d62375e7734ddc3
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@theqtcompany.com>