Add missing rate() method in java class
In the TextToSpeech android plugin, the c++ code asks through JNI
for the rate value, but the corresponding rate() method in the java class
is missing. An exception is then generated and a zero rate value is returned.
This patch adds the missing rate() method to the java source file.

Pick-to: 5.15
Change-Id: If7d1af03ed106774e33b2d50ebe84625c7bca431
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>