Wrong tts current language parsing on some Android devices
Current implementation for getting the current tts language selected
in Android gets the output of toString() to create a QLocale object.
However there is no guarantee that these calls will return two-digit
codes as required by the QLocale constructor. For example, a Samsung s10e
device returns three-digit codes ("eng" instead of "en") which causes
the instantiated QLocale object to return a generic "C" language instead
of the real locale.

This patch checks if the language and country code is two-digit or three-digit
and, in the second case, convert the three-digit code into two-digit code to
allow QLocale to be initialized correctly.

Pick-to: 5.15
Change-Id: I50163ffca57a031028e75a488d2b29ea7e9cdf4f
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>