Update commercial license headers
Updated header.COMM to the files in tqtc-qtspeech. Examples, tests
or documentation files are not updated.

The commercial license header may contain some additional
lines so that its line count equals with the earlier
license header. Reason for this is that some autotests
use hard coded line numbers and a change in the line
count causes failures in tests.

Task-number: QTQAINFRA-4221
Change-Id: I6c44137ea3a75e3b0260fee6e41f0b7afa40cdf6
Reviewed-by: Antti Kokko <antti.kokko@qt.io>