Fix build system for the Flite engine
It depends on Qt::Multimedia, so make sure we find that as an optional
component, and only enable the feature if Multimedia is found to avoid
misleading configure output (claiming that the engine is enabled, but
then it's still not getting built). Generate a warning so that users
can see what's still missing after installing flite.

The plugin code still needs to be ported to Qt 6 Multimedia APIs, so
disable the code for now via #if 0.

Change-Id: Ibb3e095dd4d2104b1ef6fb7c0410d1b285271c02
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>