Tie a QVoice to the engine it came from, add data stream operators
Only an engine can create a QVoice, so to be able to persist and load
a QVoice we need to know which engine it came from. Extend the 'data'
member to a custom struct that includes the engine name, which can then
be written to a data stream. When loading, use the same QTextToSpeech
engine and try to find a matching voice, otherwise try to find match in
the list of all available voices.

Fixes: QTBUG-50289
Change-Id: I6aaf7a2ecdded056a7f79c6691ab2e23faeef1de
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>