Flite engine: load known voice libraries dynamically
Instead of hardcoding the "kal" or "kal16" voice (which lives in
different libraries on different linux distros), try to load a list of
voice libraries dynamically.

This list is still hardcoded, and the "us" language is still hardcoded
as well, but this avoids complicating the build system to accommodate
different linux distros.

Change-Id: Iac1a043e50234fd83b7128e4f8a8e38ea9485111
Reviewed-by: Jeremy Whiting <jpwhiting@kde.org>
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>