Rewrite the WinRT implementation
Get rid of Xaml dependency, play audio through Qt Multimedia instead.
Implement an QIODevice and handlers for the asynchronous synthesize and
read operations in a single class that feeds a QAudioSink with data.

The audio format is hardcoded to 16KHz with 16bit samples and Mono.
This fits some obscure documentation and produces an audio that matches
the corresponding output through the SAPI engine.

Remove data structures that only duplicate data that we can get from
the system. While slower, none of the operations (listing available
locales and voices, getting or setting a voice) are performance critical
enough to justify the duplication.

Fixes: QTBUG-93408
Fixes: QTBUG-84213
Change-Id: I5253f316cbcced66c00b94a999b8b57b7038d88e
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>