Respect silence when pausing in the WinRT engine
If we are asked to pause, don't suspend the audio sink, as that might
stop the utterance in the middle of a word, and also in the middle of a
wave, making for both unergonomic user experience, and for unpleasant
noise when resuming.

Instead, make the audio source look for an upcoming stretch of silence,
play until the middle of that silence, and then stop providing more
data.

This makes the audio sink go into IdleState, which then translates into
QTextToSpeech::Paused. To resume, signal the audio sink that more data
is available by emitting readyRead.

This doesn't work as one would expect, so we have to explicitly suspend
and resume the sink as well. This migth be a bug in QAudioSink.

Change-Id: I767a006057a4b00742f2d56645abac9c808a5b3f
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>