Add a QML API to QTextToSpeech
As with all QML types, we need to be able to default construct, and set
properties explicitly. QTextToSpeech so far requires the setting of the
engine at construction time, and uses a "default" engine when no engine
name was set. The default engine is the arbitrarily first plugin in the
list. Turn this into a proper property. The default constructor still
instantiates the "default" engine.

The list of available locales and voices are invokable from QML as
functions.

Make QVoice a gadget with read-only properties, and register the enums.

As a drive-by, don't access the q-pointer in the private's destructor
(we are already in the QObject destructor when we get there).

Change-Id: I137cd8144ee37a08e93f3f98595baba3177b78a7
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>