Give plugins priorities
Since we support multiple engines on several platforms, and have a mock
engine for testing, we don't want to load the random first plugin in the
list as the default engine.

Intead, give each plugin a priority. Plugins based on older technology
like Windows' Speech API get a lower priority than those using newer
SDKs.

The mock engine gets a priority of -1 so it gets never selected as the
default engine (and we could also exclude it from the availableEngines
list so that it only ever used when used explicitly in a unit test).

Change-Id: I44add5cc58aeb126012afd71ce9e33c3aa04e590
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>
Reviewed-by: Jeremy Whiting <jpwhiting@kde.org>