Enable the iOS engine for all Apple platforms
The iOS engine uses AVFoundation, which has been supported on macOS
since 10.14. The regular macos engine uses the AppKit SpeechSynthesizer
framework, which has a poorer feature set, and fewer voices and
languages.

Fix the iOS plugin's build system so that it builds on macOS as well,
and give the macos engine a lower priority.

Ultimately, we can perhaps drop the macos engine completely, and rename
the iOS plugin to "darwin" or "apple" or "avfoundation".

Change-Id: Id72ee7f86d20450311e39553a2a1e0f4b8f5f0f1
Reviewed-by: Doris Verria <doris.verria@qt.io>