Documentation polish
Standardize indentation style, document properties as such, add links to
related methods, and consistently use \return and \note commands.

Document that changing the locale might change the list of available
voices, and might change the current voice. This does not fully match
the current engine implementations, but that is work in progress as part
of the porting to Qt 6.

Change-Id: I34e79f92c943c30ef10945b87372ff294864a16c
Reviewed-by: Venugopal Shivashankar <Venugopal.Shivashankar@qt.io>