Clean up unit test for QTextToSpeech
Convention in Qt is to have unit tests for a class in a subdirectory
that has the lowercase name of the class.

Follow naming convention for test functions, test basic property setters,
getters, and notify signals first, and split the more complex testing of
speech into a separate test function.

Change-Id: Ic0e0e5c6ac61949bcd39927933c85514402468bb
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>