Add unit test for QVoice
Basic test for value semantics and consistent behavior across engines.

Fix the mock engine to identify voices in a way that does not depend on
the engine instance. That's what all other engines do, and to have voices
as values persist beyond the life time of their engine, we can't use
instance pointers.

Change-Id: I4a611f484ea3c420ec3cee2a97aad721d47d4c7a
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>