Untangle voice and engine again
Amends 35769054fb172abc24b07229d025de5a732de379, which tied a voice
to the engine it came from. The engine gets its providerName set after
construction. But as QVoice instances are created by the engines
during construction, these voices won't have their engine name set.

Reading a voice from a data stream just needs to populate the public
properties, and the engine-specific data, which we already assumed
to be of a serializable type (i.e. not a pointer), which is true for
all engines.

A QVoice is still only usable by the engine that originally created
it, so if a voice is loaded from storage, then setting it on a
QTextToStream instance will fail if the engines are not the same.
This didn't change with the original change, and doens't change
with this change either. An API that allows us to find a voice that
matches certain criteria would be different anyway.

Change-Id: I9cbdb9087baef805cf2353377b026567e0f42cc3
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>