Add tests for setting voice and locale
The two properties are intertwined, changing one is very likely to
change the other: setting a locale in which the current voice is not
supported has to change voice; and setting a voice that the current
locale does not support has to change locale.

Emit signals if, and only if, the respective other attribute has been
modified after calling the engine on the setter. The engine's setter
implementation is responsible for picking a new voice or locale, and for
testing that the setter changed the value in the first place.

Adjust mock, flite, and speechd engines accordingly.

Change-Id: Iafd585b56a1a3c36b919ed6e6f29aabbd322e97c
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>