Don't emit signals when attributes don't change
Leaving it to the engine implementations to ignore and return false in
setters when the new value is the same as the old value is error prone
and unnecessary, even if engines can sometimes chose a slightly more
optimal way to compare values.

Ensure that we have the right semantics in the QTextTosSpeech class
directly, and add test coverage.

Change-Id: Ie18b9bf4577e18f9fca9f70d1310856d1ab35f7b
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>