iOS engine: compensate for gradual pitch shift, fix rate
The AVFoundation API uses a pitch multiplier that accumulates with each
utterance sent to the synthesizer. To compensate for that, keep track of
the actual pitch shift, and adjust the next multiplier accordingly.
Since we can't go below 0.5 with a multiplication factor, but might have
to go from 2.0 to 0.5 in a single shift, introduce a correction utterance
when needed that is then ignored.

For the rate calculation, we need to move up or down from the default
rate.

Fixes: QTBUG-85019
Change-Id: I06398a586f648415d69b86b2904fa09f63be9205
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>