WinRT: Fix out-of-bounds read in silence detection logic
The samples are in 16bit values, but the buffer length is in bytes. So
go only half as many steps. Once we found a stretch of silence, double
up to get the number of bytes as well.

As a drive-by:
It seems that a silence that is at the beginning of the buffer confuses
the audio sink, so only step into the middle of the silence if it does
not, otherwise send the entire block.

Change-Id: I5ee3c3d6f4f8bb007ccb861156752c87f69a1b91
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>