Add QVoice::locale property
In every engine, the locale is a property of the voice, so store it in
the voice. This makes sense for applications that want to know which
language a voice speaks, and it allows us to simplify the code for
engines that used to store the current voice and locale: one is enough.

Change-Id: I9a01ebf2d5817accf820ac5dee6a7f24524195d7
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
Reviewed-by: Jeremy Whiting <jpwhiting@kde.org>
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>