Android specific build and test fixes
Make the plugin depend on the jar target, otherwise the jar does
not get built when needed.

And since the Android engine initializes asynchronously we need
to always wait for the Ready state. This makes the tests for
properties pass; the tests waiting for the speaking to finish
abort the test run.

Fix the implementation of pitch and rate conversions to correctly
map from Qt's ranges of [-1.0, 1.0] to Android ranges of
(0.0, 2.0[, and [0.5, 2.0], respectively. Since Android operates
on floats a rate of e.g. 0.9 becomes a double of 0.89999 and the
test fails, so use string-based testing which allows us to limit
the precision.

Add a "warmup" utterance to the sayWithRates test so that the
first utterance doesn't take longer due to lazy initialization.

With these changes, the auto test now passes on Android.

Change-Id: Ie7698bcfdd60348441b1ea2d53cd2126d4355d5d
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>