Make engine specific properties configurable
For some engines it is useful to give applications more access to
engine-specific settings and properties. For instance, engines that play
the generated PCM data through QtMultimedia should have a configurable
audio device that is used for output. Other engines might have sub-
engines, like Android, that need to be selected when constructing.

These properties need to be set at construction time, using the so-far
unused QVariantMap parameter in the plugin API. Expose that through an
additional QTextToSpeech constructor and a default parameter to
QTextToSpeech::setEngine. In theory, some properties could be changeable
during the lifetime of an engine, but that makes it harder for engine
implementations without any practical benefit.

Once we have identified which properties each engine supports we can
provide documentation (the engines are currently not documented at all
anyway).

Fixes: QTBUG-63677
Fixes: QTBUG-66033
Change-Id: I4509961302bc2c5f80dca009737f47a80c221806
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>