Workaround bug in AVFoundation
The object returned as the default voice for a language code does not
include the gender flag, which breaks QVoice::operator==, as it compares
all attributes for equality.

So, if the voice we get has no gender set, ask for the same voice again
using the unique identifier. That voice object then has the gender
attribute set.

Change-Id: Ief377d7f50262f394e6808691387342a8d58d59f
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>