Documentation fixes
The name of the module is QtTextToSpeech, not Qt Speech, rename things
accordingly, esp the QML import URI, and adjust examples.

Update the module's index page to use the global macro for generating
the "Using the module" section, and add a changes file with some first
bits about porting to Qt 6.

Add an examples page and a documentation page for the Qt Quick example,
which also needs a qmake project file to prevent a qdoc warning (even
though there is no project that needs to be built).

Change-Id: I29d9126f52dc0f3a0294a27a34af89daa249db5d
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>