WinRT: Fix interaction with QAudioSink
Fixes in QtMultimedia make some of the workaround obsolete. The
audio sink now wakes up as expected from readyRead. However, we
still want to suspend it when we enter paused state so that the
sink doesn't constantly try to pull more data. But we can
resume it reliably after we resumed the source.

Don't emit readyRead if we got an empty buffer, and don't call
fetchMore synchronously from readData, as it will emit readyRead,
resulting in reentrancy. Even though the Windows API is supposed
to be asynchronous, it is synchronous as we only access in-
memory data.

Change-Id: I923a42ffd1d9242f7610f430666cc70217892b91
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>