Fix and edit the example documentation
If the exampledirs variable in qdocconf doesn't include the /speech
subdirectory, but each example page does, then the link generated to
code.qt.io is incorrect. Since the qdocconf is for the texttospeech
module, all examples will live in the same tree anyway.

Link to Qt Widgets and Qt Quick Controls from the respective examples,
which requires that those modules are added as dependencies.

Improve the phrasing of the text, and update and optimize the
screenshots of the examples.

Change-Id: I3de4e707c2aeed40735a714934adea59afdea1e8
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 86b86e5c48481f81fa43d3da49fd5177bc9925ef)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>