Fix QML type registration and documentation of Voice
Voice is a value type, so should start with a lower-case letter in QML.
The type doesn't need to be registered as it's not creatable, so the
warning from the QLM registration system showed. Register it explicitly
now, and fix the QML documentation.

And to make the enumerators in QVoice available to the QML engine,
register the Voice namespace in addition. Otherwise we cannot compare to
e.g. Voice.Female in QML code.

Change-Id: I70ebef5a167e107d15df1e29f8514465efe5d54b
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 802f2fdf4000bcab979666dd279688a0c23eb28c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>