Quick example: improve layout code
Give the text area a minimum height, it is invisible otherwise.
Remove unnecessary constraints on the root item.

The layout is still not perfect, it seems to lack margin on the
right when run on macOS, but those are probably bugs in Quick
Layouts that need to be fixed there.

Change-Id: Icd91ffb66cfec4e5a940d5a6767a8c24329c61ea
Reviewed-by: Jeremy Whiting <jpwhiting@kde.org>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 9de53a16e1abe846c7a1a63618b9bf1d20d7618a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>