Quick example: improve layout code
Give the text area a minimum height, it is invisible otherwise.
Remove unnecessary constraints on the root item.

The layout is still not perfect, it seems to lack margin on the
right when run on macOS, but those are probably bugs in Quick
Layouts that need to be fixed there.

Pick-to: 6.4
Change-Id: Icd91ffb66cfec4e5a940d5a6767a8c24329c61ea
Reviewed-by: Jeremy Whiting <jpwhiting@kde.org>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>