Android: migrate away from deprecated overload
As of API level 21, parameters have to be passed in as a Bundle.

And the onError handler in UtteranceProgressListener got a second
parameter with the error code. So overload the new version as well
and forward the old implementation.

While at it, replace the repeated string literal for the UtteranceId
with a static variable, and apply some minor code cleanup.

Change-Id: I4068240d150a34726a283d594998b42a2e9a65c0
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>