Add support for enqueuing text
Add a sayNext() slot that doesn't stop any ongoing speech, and
instead enqueues the new text. The cross-platform implementation
keeps track of enqueued texts, and processes the next text when
the engine's state changes to ready.

Make this the default behavior for synthesize(). It makes no
sense to interrupt an ongoing process, the application can just
stop an ongoing process and discard the PCM data it doesn't want.

The stateChanged signal does not get emitted when the engine's
state changes to Ready and there are texts in the queue. To
allow applications to keep track of the text that is about to be
spoken, add a new aboutToSynthesize signal that gets emitted each
time text is about to be passed down to the engine. This also
allows applications to make last-minute changes to the voice
attributes.

To accurately keep track of which text within the data structure
of the application is about to be spoken or finished, applications
do need to keep track of the text segments passed QTextToSpeech
and update their "current" iterator with the aboutToSynthesize
signal.

Task-number: QTBUG-102355
Change-Id: I7b8621e15ee8d520b156e1fd771e120ded731fd8
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>