Harden the cleanup of the synthesize callbacks
Disconnect immediately when we stop with Immediate boundary, which
includes when we destroy the engine (rather than only when the status
changes to Ready).

In the Darwin implementation, use a QPointer to detect when the engine
got destroyed. We might get a stray call to the callback block even
afterwards, which results in sporadic crashes if QTextToSpeech engines
are destroyed while synthesizing.

Change-Id: I463ed22805a8c6f2abeca3acdc3dcd3fb5ad864b
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>