Add QTextToSpeech::findVoices
QVoice is not a type that can be constructed by client code, it can only
be constructed by the engine (as the QVoice instance carries the engine
specific identifier, which an application cannot know).

Applications that want to select a voice that matches a certain (set of)
criteria - for instance "a male English voice" - have to go through some
length to first get the list of all voices (availableVoices returns only
the voices for the current locale, so one first has to go through all
locales, sets each of them, and combine the lists of voices for each
locale), and then postprocess that list to find a voice that matches.

For C++, a variadic template function returns the list of voices that
matches an arbitrary set of criteria, in an arbitrary order. Each
argument is matched against the corresponding property of a voice based
on its type. It's possible to match voices against only language or
territory as well as a fully defined QLocale object.

For QML, the function takes a map from voice's property name to
value. Since QML cannot construct a locale with an "Any" territory
or language, a "language" property is added to the voice type so that
only the language attribute of the provided locale is compared.

Make this testable by providing the mock engine with support for a
parameter "voices". That parameter takes a list of std::tuple, as
otherwise we'd have to create a VoiceData struct type of sorts that the
test can use. Extend the list of built-in voices, and improve QVoice's
debug output by printing the full locale, not just the language.

Change-Id: I5266e65932ea3db52fae92a6f50caa14dbe1f2f6
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>