Provide a VoiceSelector attached property
This makes it possible to select a voice declaratively, by specifying
the selection criteria in a grouped property. If the VoiceSelector is
specified as part of the element declaration, then the voice is selected
only once the component is completed and all selection criteria
set.

Since this adds a bunch of QML-specific code to the module, move the
existing QML binding code and module build files into a subdirectory.

Move the findVoices invokable from QTextToSpeech to the new QML-specific
subclass (it's not needed as a C++ API and was documented as internal in
QTextToSpeech).

Change-Id: I6f8907f53b513d1108f8446d57bef5975035163b
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>