API review: return an ID of the enqueued utterance
This way applications can easier keep track of things later, using the
ID as a token to identify the text. Use that ID instead of the text
string as the parameter in the aboutToSynthesize signal.

Rename the "text" parameters and variables to "utterance" to make the
relation to the BoundaryHint::Utterance clearer, and document the
Utterance enum value better (and also add the missing [since] tag).
Use a QQueue for the pendingUtterances variable instead of a list.

Change-Id: Icc2054dca52cea450736d63ca0119989b762e39a
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>