API review: remove the synthesized signal and the simple slot
They were only need for QML clients, for C++ client the version taking
a callable is more convenient. But QML clients cannot really use
QAudioFormat (there is no QML version of the type), and operating
on QByteArrays is also not something we want QML (or JavaScript code)
to do.

The engines still emit a signal, as that makes it easier to implement
the engine.

Change-Id: Ie24a41195cd5b7e27ec2b1562fb3f8e515c5adc3
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
Reviewed-by: Jeremy Whiting <jpwhiting@kde.org>