Add PauseResume capability flag
Android lacks this capability as the native TextToSpeech class has no
API for this. Since we now have the capabilities enum, add PauseResume
and don't list that for Android, so that applications can disable the
respective UI.

Do that for our examples - hide the pause & resume buttons if the
capability flag is not set.

Document the unsupported capabilities for all engines that lacks some
features

Fixes: QTBUG-113805
Change-Id: Ia8139e235f4cd968519423515e31c81285a2d349
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>