Prevent QVoice from being exported twice to QML
We need to export the type as a value type, and also export its enums
in a namespace. This resulted in duplicated entries in the qmltypes
file, which confuses Qt Design Studio.

Fix it by introducing a derived class, and using it for exporting the
enums.

Fixes: QTBUG-115363
Change-Id: Iaeb21a81ede657499a62fa5325122592ec0b203e
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit fd807f459dcf1dfbe0238686bba7eb131cd8a93d)