Examples: fix for asynchronous engine
The Android engine initializes asynchronously, so we can't assume that
the list of locales, languages, and voices is available right after
setting the engine on the QTextToSpeech object. Instead, we have to wait
for the engine to emit the stateChanged signal, and then update the list
of available options.

Fix this in both examples by factoring the code into a separate function
that we can call either immediately, or through a signal handler. In the
C++ case we can use a single-shot connection, in the QML example we
have to connect and then disconnect explicitly. In the QML example, we
also incorrectly tried to call the signal-handler of the TextToSpeech
instance as if it was a function; that doesn't work, fix that as a drive
-by.

Fixes: QTBUG-117516
Change-Id: Ibb67f00ca1fab0a080c20ba5c76a0a8765dfcdcc
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
Reviewed-by: Nicholas Bennett <nicholas.bennett@qt.io>
(cherry picked from commit 92cb62d6a1aca5785724d1616463da35c2fb71fc)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit d94733906f65df92ffddb3ded07f5bcd7434a2e7)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>