WinRT: fix hanging synthesize when RIFF header is removed
Amends 428cae98619584a51c01be135fecf6b347c117b3. If we drop the RIFF
header bytes, then we have fewer bytes available as well, so decrease
that value. Otherwise we never fetch more data.

Improve the test, which should have caught this regression, by testing
that synthesise cycles correctly through the states even if no default
audio output is present. The test was completely skipped because of
that, not detecting the regression.

Fixes: QTBUG-118668
Pick-to: 6.6
Change-Id: I3b4276c13ce2c77ec718d1b7bbca6f3421636890
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>