SAPI: fix word offsets in sayingWord signal
The event parameters take the leading XML into account, and the stored
currentText includes the XML as well, so we only have to skip the prefix
for the offset parameters of the signal.

The existing test covers this already and fails without this change,
just not CI as the Windows nodes don't have sound output devices.

Pick-to: 6.6
Change-Id: If10130131d39f65619730e13e7c8c0996e6cf0a9
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>