WinRT: don't emit Paused state if no pause was requested
If we start a synthesis process then we start the audio sink immediately
when the stream is ready. However, there might not be any data yet, so
the sink might transition briefly to suspended state, which we regularly
interpret as the text-to-speech state being paused.

Ignore such transitions if no pause was requested from the source.

This stabilizes the sayMultiple test, which without this change emits
too many state change signals and fails.

Pick-to: 6.6
Change-Id: I5158d0a0ebde6be37394cf114e1242f29df00ef2
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>