WinRT: implement boundary hint support in pause
Look for the next matching boundary, which stores its starting time.
Calculate the corresponding byte using QAudioFormat, then cut the
delivery of data off when we reach the matching byte.

In case of immediate pause, suspend the sink immediately.
The default is unchanged: we search for silence in the current chunk of
data.

Implementing boundary support for stop() might come later.

Pick-to: 6.6
Change-Id: I2b4cf06dbf8e2c55321a5364d2366c255d9e9afe
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>