Flite: immediately invoke sayingWord in flite callback
The original code tried to delay emitting sayingWord until the data is
actually processed by the QAudioSink. However the use of a timer didn't
work here, as the timer required an active event loop, while no event
loop was served.
It is less complex to immediately emit sayingWord when the initial chunk
is written to the QIODevice. Setting the ringbuffer size to 100ms, it
should be below perception.

Fixes: QTBUG-137855
Pick-to: 6.8 6.9 6.10
Change-Id: I47480571dad9813706175224192ba31afb28b70e
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>