Flite: refactor QTextToSpeechProcessorFlite
The busy waiting when accessing the QAudioSink was a bit problematic, as
we should better go through the event loop. Flite doesn't seem to
provide any nice "pull" API, so we move the flite real-time synthesis to
a worker thread.
This worker thread produces the tokenization as well as the PCM data,
which is fed to the QTextToSpeechProcessorFlite and its event loop.
There we feed the QAudioSink, handle pause/resume and emit sayingWord
signals.

Task-number: QTBUG-137855
Change-Id: I65265c294a908cb6ae95afab0a0def9408b46859
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>